Made JSON.stringify comptible with javascript JSON
The current JSON.stringify method escaped characters in a way that
was NOT compatible with the current javascript JSON implementation.
For example, the string single-character "\n" was being translated
into:

"\"\
\""

Instead of being stringifyed as: "\"\n\"". This is because the
character \n was being added, instead of adding an "n" preceeded
by an "\\".

This commit fixes such problem by using a direct mapping from
escapable characters into their escaped stringified version.